回り込み攻撃仮完成、加速度とエフェクトを付けてみる１

+++ アクションゲーム/EffectBase.h
@@ -40,7 +40,7 @@ public:
 	void BaseUpdate();//‹¤’ÊXVˆ—
 
 	void ResetLife() { m_lifeframe = 0; m_live = true; }
-	void SetLifeFlame(float flame) { m_lifeframe = flame; }
+	void SetLifeFlame(int flame) { m_lifeframe = flame; }
 
 
 	/*void PlayEffect(
